Add example of a deeplink in the demo app
